refactor: extract loadTicket and runResearchWorkflow to shared module
Moves remaining duplicated code (loadTicket, supervisor init, run loop)
into display-helpers.ts. Entry points now only define their unique
getTicketPath logic and help hints.
